Add SceneGeometry implementation and update PoseLib port
- Introduced SceneGeometry.cpp to handle geometric and KD-tree-based scene methods, including point cloud normal estimation, sparse surface estimation, and ROI cropping.
- Disabled tests in PoseLib port to enforce static linkage on Windows builds.
- Bumped PoseLib port version to 2 for the latest changes and unified API for camera pose estimation.
