+++ CMakeLists.txt
@@ -127,10 +127,10 @@ if(OpenMVS_ENABLE_IPO)
 	INCLUDE(CheckIPOSupported)
 	check_ipo_supported(RESULT IPO_SUPPORTED OUTPUT IPO_ERROR)
 	if(IPO_SUPPORTED)
-		SET_PROPERTY(GLOBAL PROPERTY CMAKE_INTERPROCEDURAL_OPTIMIZATION TRUE)
-		MESSAGE(STATUS "Interprocedural optimization enabled")
+		MESSAGE(STATUS "Interprocedural optimization enabled (MSVC: /GL + /LTCG in Release)")
 	else()
 		MESSAGE(WARNING "Interprocedural optimization is not supported: ${IPO_ERROR}")
+		SET(OpenMVS_ENABLE_IPO OFF CACHE BOOL "Whether to enable interprocedural optimization" FORCE)
 	endif()
 endif()
